Organize test js files
